math.h works now, including fix to the inl files

+++ include/aws/common/math.gcc_overflow.inl
@@ -105,7 +105,7 @@ AWS_STATIC_IMPL int aws_add_u32_checked(uint32_t a, uint32_t b, uint32_t *r) {
 /**
  * Adds a + b. If the result overflows, returns 2^32 - 1.
  */
-AWS_STATIC_IMPL uint64_t aws_add_u32_saturating(uint32_t a, uint32_t b) {
+AWS_STATIC_IMPL uint32_t aws_add_u32_saturating(uint32_t a, uint32_t b) {
     uint32_t res;
 
     if (__builtin_add_overflow(a, b, &res)) {
