+++ include/aws/common/math.gcc_x64_asm.inl
@@ -165,7 +165,7 @@ AWS_STATIC_IMPL int aws_add_u32_checked(uint32_t a, uint32_t b, uint32_t *r) {
 /**
  * Adds a + b. If the result overflows, returns 2^32 - 1.
  */
-AWS_STATIC_IMPL uint64_t aws_add_u32_saturating(uint32_t a, uint32_t b) {
+AWS_STATIC_IMPL uint32_t aws_add_u32_saturating(uint32_t a, uint32_t b) {
     /* We can use inline assembly to do this efficiently on x86-64 and x86. */
 
     __asm__("addl %[arg1], %[arg2]\n" /* [arga] = [arga] + [argb] */
